billing/invoice: rename internal package to pkg.
