+++ systems/billing/invoice/Makefile
@@ -22,7 +22,7 @@ deps:
 gen:
 	protoc -I ./pb -I ../../common/pb -I ../../common/pb/dep --go_out=. --go-grpc_out=. --go-grpc_opt=require_unimplemented_servers=true --govalidators_out=. \
 		invoice.proto
-	mockery --all --recursive --dir ./internal
+	mockery --all --recursive --dir ./pkg
 	mockery --dir ./pb  --all --recursive --output ./pb/gen/mocks
 
 clean:
