+++ systems/billing/invoice/pkg/config.go
@@ -1,4 +1,4 @@
-package internal
+package pkg
 
 import (
 	"time"
